Move CredentialDeviceType into structs
