+++ webauthn/helpers/structs.py
@@ -56,6 +56,7 @@ class WebAuthnBaseModel(BaseModel):
             # values
             return v
 
+
 ################
 #
 # Fundamental data structures
@@ -545,3 +546,24 @@ class AuthenticationCredential(WebAuthnBaseModel):
     type: Literal[
         PublicKeyCredentialType.PUBLIC_KEY
     ] = PublicKeyCredentialType.PUBLIC_KEY
+
+
+################
+#
+# Credential Backup State
+#
+################
+
+
+class CredentialDeviceType(str, Enum):
+    """A determination of the number of devices a credential can be used from
+
+    Members:
+        `SINGLE_DEVICE`: A credential that is bound to a single device
+        `MULTI_DEVICE`: A credential that can be used from multiple devices (e.g. passkeys)
+
+    https://w3c.github.io/webauthn/#sctn-credential-backup (L3 Draft)
+    """
+
+    SINGLE_DEVICE = "single_device"
+    MULTI_DEVICE = "multi_device"
